:white_check_mark: Add unit test for purify_html().
In preparation to update HTMLPurifier library add a unit test.
Please add more HTML that you need to have working and we can check
against regressions.
